Converting to a Dagster resource
