+++ saf_gdelt.egg-info/top_level.txt
@@ -1,0 +1,1 @@
+saf_gdelt
