+++ setup.py
@@ -1,10 +1,10 @@
 from setuptools import find_packages, setup
 
 setup(
-    name='dpf_gdelt',
-    packages=find_packages(include=['dpf_gdelt']),
+    name='saf_gdelt',
+    packages=find_packages(include=['saf_gdelt']),
     version='0.1.0',
-    description='GDELT mining package for the Lantrns Analytics\' data product framework',
+    description='GDELT resource library for the Social Analytics Framework',
     author='Olivier Dupuis',
     license='MIT',
 )
